+++ labs/dx2/Form1.cs
@@ -33,7 +33,7 @@ namespace dx
 
 
             device.RenderState.CullMode = Cull.CounterClockwise;
-            device.Transform.Projection = Matrix.PerspectiveFovLH((float)Math.PI / 2, this.Width / this.Height, 1f, 50f);
+            device.Transform.Projection = Matrix.PerspectiveFovLH((float)Math.PI / 2, (float)this.Width / this.Height, 1f, 50f);
 
 
             device.RenderState.Lighting = true;
